+++ CHANGELOG.md
@@ -3,13 +3,15 @@
 ### Fix
 
 - fix metamask error on cancelling send and fix bug when disconnecting through metamask [#5337](https://github.com/MyEtherWallet/MyEtherWallet/pull/5337)
-- handle shit token send error [#5330](https://github.com/MyEtherWallet/MyEtherWallet/pull/5330)
+- handle crap token send error [#5330](https://github.com/MyEtherWallet/MyEtherWallet/pull/5330)
 - Add filter low fiat balance tokens [#5333](https://github.com/MyEtherWallet/MyEtherWallet/pull/5333)
+- fix font loading [#5356](https://github.com/MyEtherWallet/MyEtherWallet/pull/5356)
 
 # Feat
 
 - Add blinking info for approval on ledger or trezor [#5331](https://github.com/MyEtherWallet/MyEtherWallet/pull/5331)
 - store selected pair in swap and trade, and set it once user navigates back [#5335](https://github.com/MyEtherWallet/MyEtherWallet/pull/5335)
+- store chosen network on bridge [#5343](https://github.com/MyEtherWallet/MyEtherWallet/pull/5343)
 - fix error when watchlist tokens are cleared [#5336](https://github.com/MyEtherWallet/MyEtherWallet/pull/5336)
 
 ### v7.0.2